python/requirements.txt
numpy
scipy
pillow
transformers
torch
accelerate
librosa
soundfile
pandas
matplotlib
